Refactor manager and cluster into SDK and core
