+++ api/api/controllers/cluster_controller.py
@@ -3,22 +3,20 @@
 # This program is a free software; you can redistribute it and/or modify it under the terms of GPLv2
 
 import asyncio
-import connexion
 import datetime
 import logging
 
-from api.models.base_model_ import Data
-from api.util import remove_nones_to_dict, exception_handler, parse_api_param, raise_if_exc
+import connexion
+
 import wazuh.cluster as cluster
-import wazuh.core.cluster.control as cluster_control
-import wazuh.configuration as configuration
-from wazuh import Wazuh
-from wazuh.core.cluster.dapi.dapi import DistributedAPI, set_system_nodes
-from wazuh.exception import WazuhError
+import wazuh.common as common
 import wazuh.manager as manager
 import wazuh.stats as stats
 from api.authentication import get_permissions
-import wazuh.common as common
+from api.models.base_model_ import Data
+from api.util import remove_nones_to_dict, exception_handler, parse_api_param, raise_if_exc
+from wazuh.core.cluster.dapi.dapi import DistributedAPI, get_system_nodes
+from wazuh.exception import WazuhError
 
 loop = asyncio.get_event_loop()
 logger = logging.getLogger('wazuh')
@@ -26,15 +24,14 @@ logger = logging.getLogger('wazuh')
 
 @exception_handler
 def get_cluster_node(pretty=False, wait_for_complete=False):
-    """Get information about the local node.
-
-    Returns basic information about the cluster node receiving the request.
+    """Get basic information about the local node.
 
     :param pretty: Show results in human-readable format
     :param wait_for_complete: Disable timeout response
     """
     f_kwargs = {}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=cluster.get_node_wrapper,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='local_any',
@@ -42,9 +39,9 @@ def get_cluster_node(pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -76,6 +73,7 @@ def get_cluster_nodes(pretty=False, wait_for_complete=False, offset=0, limit=Non
                 'select': select,
                 'filter_type': type_}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=cluster.get_nodes_info,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='local_master',
@@ -84,9 +82,9 @@ def get_cluster_nodes(pretty=False, wait_for_complete=False, offset=0, limit=Non
                           pretty=pretty,
                           logger=logger,
                           local_client_arg='lc',
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -106,6 +104,7 @@ def get_healthcheck(pretty=False, wait_for_complete=False, list_nodes=None):
     """
     f_kwargs = {'filter_node': list_nodes}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=cluster.get_health_nodes,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='local_master',
@@ -114,9 +113,9 @@ def get_healthcheck(pretty=False, wait_for_complete=False, list_nodes=None):
                           pretty=pretty,
                           logger=logger,
                           local_client_arg='lc',
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -131,6 +130,7 @@ def get_status(pretty=False, wait_for_complete=False):
     """
     f_kwargs = {}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=cluster.get_status_json,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='local_master',
@@ -138,9 +138,9 @@ def get_status(pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
     response = Data(data)
 
@@ -156,6 +156,7 @@ def get_config(pretty=False, wait_for_complete=False):
     """
     f_kwargs = {}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=cluster.read_config_wrapper,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='local_any',
@@ -163,9 +164,9 @@ def get_config(pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -173,9 +174,7 @@ def get_config(pretty=False, wait_for_complete=False):
 
 @exception_handler
 def get_status_node(node_id, pretty=False, wait_for_complete=False):
-    """Get a specified node's status 
-
-    Returns the status of all Wazuh daemons in node node_id
+    """Get a specified node's Wazuh daemons status
 
     :param node_id: Cluster node name.
     :param pretty: Show results in human-readable format
@@ -183,6 +182,7 @@ def get_status_node(node_id, pretty=False, wait_for_complete=False):
     """
     f_kwargs = {'node_id': node_id}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.get_status,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -190,9 +190,9 @@ def get_status_node(node_id, pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -210,6 +210,7 @@ def get_info_node(node_id, pretty=False, wait_for_complete=False):
     """
     f_kwargs = {'node_id': node_id}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.get_basic_info,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -217,9 +218,9 @@ def get_info_node(node_id, pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -227,9 +228,7 @@ def get_info_node(node_id, pretty=False, wait_for_complete=False):
 
 @exception_handler
 def get_configuration_node(node_id, pretty=False, wait_for_complete=False, section=None, field=None):
-    """Get a specified node's configuration 
-
-    Returns wazuh configuration used in node {node_id}
+    """Get a specified node's configuration (ossec.conf)
 
     :param node_id: Cluster node name.
     :param pretty: Show results in human-readable format
@@ -241,6 +240,7 @@ def get_configuration_node(node_id, pretty=False, wait_for_complete=False, secti
                 'section': section,
                 'field': field}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.read_ossec_conf,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -248,9 +248,9 @@ def get_configuration_node(node_id, pretty=False, wait_for_complete=False, secti
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -281,6 +281,7 @@ def get_stats_node(node_id, pretty=False, wait_for_complete=False, date=None):
                 'day': day,
                 'date': True if date else False}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=stats.totals,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -288,9 +289,9 @@ def get_stats_node(node_id, pretty=False, wait_for_complete=False, date=None):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -309,6 +310,7 @@ def get_stats_hourly_node(node_id, pretty=False, wait_for_complete=False):
     """
     f_kwargs = {'node_id': node_id}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=stats.hourly,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -316,9 +318,9 @@ def get_stats_hourly_node(node_id, pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
     response = Data(data)
 
@@ -338,6 +340,7 @@ def get_stats_weekly_node(node_id, pretty=False, wait_for_complete=False):
     """
     f_kwargs = {'node_id': node_id}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=stats.weekly,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -345,9 +348,9 @@ def get_stats_weekly_node(node_id, pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
     response = Data(data)
 
@@ -365,6 +368,7 @@ def get_stats_analysisd_node(node_id, pretty=False, wait_for_complete=False):
     f_kwargs = {'node_id': node_id,
                 'filename': common.analysisd_stats}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=stats.get_daemons_stats,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -372,9 +376,9 @@ def get_stats_analysisd_node(node_id, pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
     response = Data(data)
 
@@ -392,6 +396,7 @@ def get_stats_remoted_node(node_id, pretty=False, wait_for_complete=False):
     f_kwargs = {'node_id': node_id,
                 'filename': common.remoted_stats}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=stats.get_daemons_stats,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -399,9 +404,9 @@ def get_stats_remoted_node(node_id, pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
     response = Data(data)
 
@@ -436,6 +441,7 @@ def get_log_node(node_id, pretty=False, wait_for_complete=False, offset=0, limit
                 'category': category,
                 'type_log': type_log}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.ossec_log,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -443,9 +449,9 @@ def get_log_node(node_id, pretty=False, wait_for_complete=False, offset=0, limit
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -461,6 +467,7 @@ def get_log_summary_node(node_id, pretty=False, wait_for_complete=False):
     """
     f_kwargs = {'node_id': node_id}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.ossec_log_summary,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -468,9 +475,9 @@ def get_log_summary_node(node_id, pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -488,6 +495,7 @@ def get_files_node(node_id, path, pretty=False, wait_for_complete=False):
     f_kwargs = {'node_id': node_id,
                 'path': path}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.get_file,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -495,9 +503,9 @@ def get_files_node(node_id, path, pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -528,6 +536,7 @@ def put_files_node(body, node_id, path, overwrite=False, pretty=False, wait_for_
                 'overwrite': overwrite,
                 'content': body}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.upload_file,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -535,9 +544,9 @@ def put_files_node(body, node_id, path, overwrite=False, pretty=False, wait_for_
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -555,6 +564,7 @@ def delete_files_node(node_id, path, pretty=False, wait_for_complete=False):
     f_kwargs = {'node_id': node_id,
                 'path': path}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.delete_file,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -562,9 +572,9 @@ def delete_files_node(node_id, path, pretty=False, wait_for_complete=False):
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -580,6 +590,7 @@ def put_restart(pretty=False, wait_for_complete=False, list_nodes='*'):
     """
     f_kwargs = {'node_list': list_nodes}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.restart,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -588,9 +599,9 @@ def put_restart(pretty=False, wait_for_complete=False, list_nodes='*'):
                           pretty=pretty,
                           logger=logger,
                           broadcasting=list_nodes == '*',
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -607,6 +618,7 @@ def get_conf_validation(pretty=False, wait_for_complete=False, list_nodes='*'):
     """
     f_kwargs = {'node_list': list_nodes}
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.validation,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -615,9 +627,9 @@ def get_conf_validation(pretty=False, wait_for_complete=False, list_nodes='*'):
                           pretty=pretty,
                           logger=logger,
                           broadcasting=list_nodes == '*',
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
@@ -637,6 +649,7 @@ def get_node_config(node_id, component, wait_for_complete=False, pretty=False, *
                 'config': kwargs.get('configuration', None)
                 }
 
+    nodes = loop.run_until_complete(get_system_nodes())
     dapi = DistributedAPI(f=manager.get_config,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
@@ -644,9 +657,9 @@ def get_node_config(node_id, component, wait_for_complete=False, pretty=False, *
                           wait_for_complete=wait_for_complete,
                           pretty=pretty,
                           logger=logger,
-                          rbac_permissions=get_permissions(connexion.request.headers['Authorization'])
+                          rbac_permissions=get_permissions(connexion.request.headers['Authorization']),
+                          nodes=nodes
                           )
-    loop.run_until_complete(set_system_nodes())
     data = raise_if_exc(loop.run_until_complete(dapi.distribute_function()))
 
     return data, 200
