+++ framework/wazuh/core/cluster/tests/test_cluster.py
@@ -6,6 +6,7 @@ import time
 from datetime import datetime
 from unittest.mock import patch, mock_open
 
+import wazuh.core.cluster.cluster
 import wazuh.core.cluster.utils
 
 with patch('wazuh.common.ossec_uid'):
@@ -101,7 +102,7 @@ def test_checking_configuration(read_config):
         m.return_value = read_config.copy()
         with pytest.raises(WazuhException, match=r'.* 3004 .*'):
             configuration = wazuh.core.cluster.utils.read_config()
-            cluster.check_cluster_config(configuration)
+            wazuh.core.cluster.cluster.check_cluster_config(configuration)
 
 
 agent_info = b"""Linux |agent1 |3.10.0-862.el7.x86_64 |#1 SMP Fri Apr 20 16:44:24 UTC 2018 |x86_64 [CentOS Linux|centos: 7 (Core)] - Wazuh v3.7.2 / d10d46b48c280384e8773a5fa24ecacb
@@ -123,7 +124,7 @@ def test_merge_agent_info(stat_mock, listdir_mock):
     stat_mock.return_value.st_size = len(agent_info)
 
     with patch('builtins.open', mock_open(read_data=agent_info)) as m:
-        cluster.merge_agent_info('agent-info', 'worker1')
+        wazuh.core.cluster.cluster.merge_agent_info('agent-info', 'worker1')
         m.assert_any_call(common.ossec_path + '/queue/cluster/worker1/agent-info.merged', 'wb')
         m.assert_any_call(common.ossec_path + '/queue/agent-info/agent1-any', 'rb')
         m.assert_any_call(common.ossec_path + '/queue/agent-info/agent2-any', 'rb')
@@ -140,5 +141,6 @@ def test_merge_agent_info(stat_mock, listdir_mock):
 def test_unmerge_agent_info(stat_mock, agent_info, exception):
     stat_mock.return_value.st_size = len(agent_info)
     with patch('builtins.open', mock_open(read_data=agent_info)) as m:
-        agent_infos = list(cluster.unmerge_agent_info('agent-info', '/random/path', 'agent-info.merged'))
+        agent_infos = list(
+            wazuh.core.cluster.cluster.unmerge_agent_info('agent-info', '/random/path', 'agent-info.merged'))
         assert len(agent_infos) == (1 if exception is None else 0)
