+++ framework/wazuh/core/cluster/utils.py
@@ -13,8 +13,7 @@ from glob import glob
 from operator import setitem
 from os.path import join, exists
 
-import wazuh.common as common
-from wazuh import common, WazuhException
+from wazuh import common
 from wazuh.configuration import get_ossec_conf
 from wazuh.exception import WazuhException, WazuhError, WazuhInternalError
 from wazuh.results import WazuhResult
